requirements.txt
attrs==25.1.0
aws-cdk-lib==2.181.1
aws-cdk.asset-awscli-v1==2.2.226
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==39.2.20
aws_lambda_powertools==3.7.0
boto3==1.37.6
botocore==1.37.6
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
constructs==10.4.2
idna==3.10
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.108.0
publication==0.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.3.0
